[JavaScriptEngine] Update strings to behave closer to the spec
As per: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String#string_primitives_and_string_objects

String literals are no longer auto converted to a StringObject. Instead we introduced
a new interpretAsObject method that will convert string primitives to a StringObject to
allow us to call methods on them as if they were an object. We use this in contexts
where we expect to operate on an object such as member/index access.